backend/requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Data Processing
pandas==2.1.3
numpy<2.0.0

# Environment Configuration
python-dotenv==1.0.0

# LangChain - Unified LLM Framework
langchain>=0.1.0
langchain-core>=0.1.0

# LangChain LLM Providers (install based on your API keys)
langchain-google-genai>=0.0.6    # For Google Gemini
# langchain-openai>=0.0.5        # For OpenAI GPT (uncomment if needed)
# langchain-anthropic>=0.1.1     # For Claude (uncomment if needed)
# langchain-groq>=0.0.1          # For Groq (uncomment if needed)
